--- src/types/TemplateTypes.tsx
@@ -1,13 +1,0 @@
-type Pins = {
-	id: string
-	x: Number
-	y: Number
-}
-
-export type TemplateTypes = {
-	id: string
-	name: string
-	image: string
-	pins: Pins[]
-	createdBy: string
-}
